refactor(ui): move ValidationBadge and ValidationTooltip into the alert family
