+++ js/app/src/components/core/alert/ValidationBadge.tsx
@@ -1,13 +1,8 @@
 import { css, keyframes } from "@emotion/react";
-import type { ReactNode } from "react";
 import { Pressable } from "react-aria";
 
-import { getSeverityIcon } from "@phoenix/components/core/alert/getSeverityIcon";
-import { Text } from "@phoenix/components/core/content";
-import { Flex } from "@phoenix/components/core/layout";
-import { Tooltip } from "@phoenix/components/core/tooltip";
-import type { TooltipProps } from "@phoenix/components/core/tooltip";
-import type { SeverityLevel } from "@phoenix/components/core/types";
+import type { SeverityLevel } from "../types";
+import { getSeverityIcon } from "./getSeverityIcon";
 
 /**
  * Grows a badge out from its leading edge. Animating max-width alongside
@@ -106,41 +101,5 @@ export function ValidationBadge({
         <span className="validation-badge__message">{children}</span>
       </div>
     </Pressable>
-  );
-}
-
-export type ValidationTooltipProps = Omit<TooltipProps, "children"> & {
-  /** Heading of the tooltip. */
-  title: string;
-  /** Detail rendered below the title. */
-  children?: ReactNode;
-  /** @default "danger" */
-  variant?: SeverityLevel;
-};
-
-/**
- * The tooltip half of a validation status: the severity icon beside a heavy
- * title with the detail below. Every other `Tooltip` prop passes through, so
- * a composition can place and style it like any tooltip.
- */
-export function ValidationTooltip({
-  title,
-  children,
-  variant = "danger",
-  placement = "top",
-  ...tooltipProps
-}: ValidationTooltipProps) {
-  return (
-    <Tooltip placement={placement} {...tooltipProps}>
-      <Flex direction="row" gap="size-100" alignItems="start">
-        {getSeverityIcon(variant, { filled: false })}
-        <Flex direction="column" gap="size-25">
-          <Text size="S" weight="heavy">
-            {title}
-          </Text>
-          {children}
-        </Flex>
-      </Flex>
-    </Tooltip>
   );
 }
