+++ js/app/src/components/core/alert/index.tsx
@@ -1,1 +1,3 @@
 export * from "./Alert";
+export * from "./ValidationBadge";
+export * from "./ValidationTooltip";
