+++ js/app/src/components/core/field/index.tsx
@@ -12,4 +12,3 @@ export * from "./DebouncedSearch";
 export * from "./FieldDangerIcon";
 export * from "./FieldSuccessIcon";
 export { TextArea } from "react-aria-components";
-export * from "./ValidationBadge";
